Refactor the helper functions
